+++ compiler_opt/rl/data_collector.py
@@ -19,7 +19,7 @@ import time
 from typing import Dict, Iterator, Tuple, Sequence
 
 import numpy as np
-from compiler_opt.rl import compilation_runner
+from compiler_opt.rl import policy_saver
 from tf_agents.trajectories import trajectory
 
 # Deadline for data collection.
@@ -52,7 +52,7 @@ class DataCollector(metaclass=abc.ABCMeta):
 
   @abc.abstractmethod
   def collect_data(
-      self, policy: compilation_runner.Policy
+      self, policy: policy_saver.Policy
   ) -> Tuple[Iterator[trajectory.Trajectory], Dict[str, Dict[str, float]]]:
     """Collect data for a given policy.
